feat(training): BF16 defaults + CI applies the ggml patch
CI: all five build jobs (release x3, cache-warm x2) gain an "Apply
engine patches" step for engine/patches/bf16-out-prod.patch, placed
AFTER "Restore source file timestamps" so the patched files carry
fresh mtimes and their cached objects rebuild — patching before the
restore would revive the v1.1.3 stale-object bug class.

Defaults: DiT --mirror and LM --weights now default to bf16 at the
server + UI tiers (bare-CLI engine defaults stay f32). Both trainers
degrade gracefully instead of dying when bf16 cannot apply: DiT falls
back to the f32 mirror on non-CUDA backends (was fatal); LM falls back
to f32-window on non-CUDA backends (prevents the documented CPU
out_prod GGML_ABORT) or on a non-BF16-native base (was a hard fatal),
and every JSONL/log field reports the mode actually used.

+++ ui/src/components/training-studio/TrainDitForm.tsx
@@ -121,8 +121,10 @@ export const TRAIN_DIT_DEFAULTS: TrainDitFormState = {
   milestoneStep: 0.1,
   milestoneKeep: 6,
   vramReserveMb: 2048,
-  // Stays 'f32' for both adapter types until BF16 is ear-validated.
-  mirror: 'f32',
+  // Default 'bf16' for both adapter types (2026-07-29); TRAIN_DIT_LOKR_DEFAULTS
+  // spreads this object and doesn't override mirror, so LoKR inherits it too.
+  // The engine itself falls back to 'f32' with a warning on a non-CUDA backend.
+  mirror: 'bf16',
   stages: ['train', 'export'],
   overwrite: false,
   stopEngine: true,
